Refactor project structure and reorganize components
