+++ src/App.tsx
@@ -4,7 +4,7 @@ import { QueryClientProvider } from "@tanstack/react-query";
 import { Toaster } from "@/components/ui/toaster";
 import { TooltipProvider } from "@/components/ui/tooltip";
 import Home from "@/pages/Home";
-import NotFound from "@/pages/not-found";
+import NotFound from "@/pages/NotFound";
 import OurServices from "@/pages/OurServices";
 
 function Router() {
