+++ src/components/products/CategoryNavbar.tsx
