+++ src/components/qr-code/index.tsx
