+++ src/pages/Home/index.tsx
@@ -1,14 +1,14 @@
   import { useState, useEffect } from "react";
-  import { Header } from "@/components/Header";
-  import { CategoryNavbar } from "@/components/CategoryNavbar";
-  import { ProductGrid } from "@/components/ProductGrid";
-  import { LoadingState } from "@/components/LoadingState";
-  import { EmptyState } from "@/components/EmptyState";
-  import { ErrorState } from "@/components/ErrorState";
-  import { Pagination } from "@/components/Pagination";
-  import { Footer } from "@/components/Footer";
-  import { CustomAdsLeft } from "@/components/CustomAdsLeft";
-  import { AdBannerVertical } from "@/components/AdBannerVertical";
+  import { Header } from "@/components/layout/Header";
+  import { CategoryNavbar } from "@/components/products/CategoryNavbar";
+  import { ProductGrid } from "@/components/products/ProductGrid";
+  import { LoadingState } from "@/components/common/LoadingState";
+  import { EmptyState } from "@/components/common/EmptyState";
+  import { ErrorState } from "@/components/common/ErrorState";
+  import { Pagination } from "@/components/common/Pagination";
+  import { Footer } from "@/components/layout/Footer";
+  import { CustomAdsLeft } from "@/components/ads/CustomAdsLeft";
+  import { AdBannerVertical } from "@/components/ads/AdBannerVertical";
   import { useFetchData } from "@/hooks/useFetchData";
   import type { Category } from "@shared/schema";
